requirements.txt
numpy==1.21.0
pandas==1.3.2
matplotlib==3.4.3
scikit-learn==0.24.2
tensorflow==2.6.0
GitPython
packaging
